refactor: theme file 정리, 안쓰는 변수들 제거
