+++ client/src/styles/styles.ts
@@ -1,0 +1,21 @@
+const global = {
+  'html, body': {
+    h: 'full',
+    m: '0',
+    p: '0',
+    fontFamily: 'Pretendard',
+  },
+  // footer 바닥에 붙어있게 하기 위한 css
+  '#__next': {
+    h: 'full',
+    display: 'flex',
+    flexDirection: 'column',
+    fontFamily: 'Pretendard',
+  },
+};
+
+const styles = {
+  global,
+};
+
+export default styles;
